config/requirements/base.txt
django-model-utils==4.0.0
pyyaml==5.2
psycopg2==2.8.4
